requirements.txt
flask==2.0.2
kubernetes==21.7.0
pyyaml==6.0
gunicorn==20.1.0